perubahan terjadi

+++ routes/web.php
@@ -1,10 +1,11 @@
 <?php
 
 use App\Models\DataParking;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\LoginController;
 use App\Http\Controllers\DataParkingController;
-use Illuminate\Support\Facades\Auth;
+use App\Http\Controllers\LandingPageController;
 
 /*
 |--------------------------------------------------------------------------
@@ -16,17 +17,27 @@ use Illuminate\Support\Facades\Auth;
 | contains the "web" middleware group. Now create something great!
 |
 */
-// Login Controller
-Route::get('/', [LoginController::class, 'index'])->middleware('guest')->name('login');
-Route::get('/login', [LoginController::class, 'index'])->middleware('guest')->name('login');
-Route::post('/login', [LoginController::class, 'postlogin']);
-Route::post('/login/logout', [LoginController::class, 'logout']);
-Route::get('/login/logout', [LoginController::class, 'index']);
+/*landing page */
+
+Route::get('/', [LandingPageController::class, 'index']);
+
+
+
+/* Register  */
+
+Route::get('/admin/register', [LoginController::class, 'register'])->middleware('guest')->name('login');
+Route::post('/admin/register', [LoginController::class, 'postregister'])->middleware('guest')->name('login');   
+
+// Login Controller 
+Route::get('/admin/login', [LoginController::class, 'index'])->middleware('guest')->name('login');
+Route::post('/admin/login', [LoginController::class, 'postlogin'])->middleware('guest')->name('login');
+Route::post('/admin/login/logout', [LoginController::class, 'logout']);
+Route::get('/admin/login/logout', [LoginController::class, 'logout']);
 
 
 /* eparking controller */
-Route::get('/eparking',[DataParkingController::class, 'index'])->middleware('auth');
-Route::post('/eparking', [DataParkingController::class, 'store'])->middleware('auth');
-Route::delete('/eparking/{id}', [DataParkingController::class, 'destroy'])->middleware('auth');
-Route::put('/eparking/{id}', [DataParkingController::class, 'show'])->middleware('auth');
-Route::get('/eparking/{id}', [DataParkingController::class, 'show'])->middleware('auth');
+Route::get('/admin/eparking',[DataParkingController::class, 'index'])->middleware('auth');
+Route::post('/admin/eparking', [DataParkingController::class, 'store'])->middleware('auth');
+Route::delete('/admin/eparking/{id}', [DataParkingController::class, 'destroy'])->middleware('auth');
+Route::put('/admin/eparking/{id}', [DataParkingController::class, 'show'])->middleware('auth');
+Route::get('/admin/eparking/{id}', [DataParkingController::class, 'show'])->middleware('auth');
